move fi and fuzzy result savers to services
