+++ biofefi/feature_importance/interpreter.py
@@ -2,7 +2,7 @@ import os
 
 import pandas as pd
 
-from biofefi.feature_importance.call_methods import save_importance_results
+from biofefi.services.feature_importance.results import save_importance_results
 from biofefi.services.feature_importance.ensemble_methods import (
     calculate_ensemble_majorityvote,
     calculate_ensemble_mean,
